+++ visualizers/SphereShadow.cpp
@@ -15,32 +15,32 @@ using namespace rt;
 
 int main()
 {
-    color black(0.0f, 0.0f, 0.0f);
-    color red(1.0f, 0.0f, 0.0f);
-    color blue(0.0f, 0.0f, 1.0f);
-    color yellow(1.0f, 1.0f, 0.0f);
+    color black(0.0, 0.0, 0.0);
+    color red(1.0, 0.0, 0.0);
+    color blue(0.0, 0.0, 1.0);
+    color yellow(1.0, 1.0, 0.0);
 
     constexpr size_t CANVAS_SIZE = 250;
     canvas c(CANVAS_SIZE, CANVAS_SIZE, black);
 
-    point ray_origin = point(0.0f, 0.0f, -10.0f);
-    float wall_z = 10.0f;
-    float wall_size = 7.0f;
+    point ray_origin = point(0.0, 0.0, -10.0);
+    double wall_z = 10.0;
+    double wall_size = 7.0;
 
-    float pixel_size = wall_size / CANVAS_SIZE;
-    float half = wall_size / 2.0f;
+    double pixel_size = wall_size / CANVAS_SIZE;
+    double half = wall_size / 2.0;
 
     sphere s1 = sphere();
     sphere s2 = sphere();
     sphere s3 = sphere();
 
-    matrix<4> transform_s1 = translation(0.0f, 0.0f, 0.0f);
+    matrix<4> transform_s1 = translation(0.0, 0.0, 0.0);
     s1.set_transform(transform_s1);
 
-    matrix<4> transform_s2 = translation(0.5f, 0.5f, -0.5f) * scale(0.5f, 0.5f, 0.5f);
+    matrix<4> transform_s2 = translation(0.5, 0.5, -0.5) * scale(0.5, 0.5, 0.5);
     s2.set_transform(transform_s2);
 
-    matrix<4> transform_s3 = translation(-0.5f, -0.5f, -0.5f) * scale(0.5f, 0.5f, 0.5f);
+    matrix<4> transform_s3 = translation(-0.5, -0.5, -0.5) * scale(0.5, 0.5, 0.5);
     s3.set_transform(transform_s3);
 
     // Iterate over every row (y) and column (x) of the canvas
@@ -48,12 +48,12 @@ int main()
     for (size_t y = 0; y < CANVAS_SIZE; ++y)
     {
         // Compute the world y coordinate (top = +half, bottom = -half)
-        float world_y = half - pixel_size * y;
+        double world_y = half - pixel_size * y;
 
         for (size_t x = 0; x < CANVAS_SIZE; ++x)
         {
             // Compute the world x coordinate (left = -half, right = +half)
-            float world_x = -half + pixel_size * x;
+            double world_x = -half + pixel_size * x;
 
             // Define the point on the wall that the ray will target
             point position = point(world_x, world_y, wall_z);
